+++ base/rico-client/src/main/java/dev/rico/internal/client/http/ClientHttpExecutorImpl.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2018 Karakun AG.
+ * Copyright 2015-2018 Canoo Engineering AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package dev.rico.internal.client.http;
+
+import dev.rico.internal.core.http.HttpExecutorImpl;
+import dev.rico.internal.core.http.HttpProvider;
+import dev.rico.internal.core.Assert;
+import dev.rico.client.ClientConfiguration;
+import dev.rico.core.http.HttpExecutor;
+import dev.rico.core.http.BadResponseException;
+import dev.rico.core.http.HttpException;
+import dev.rico.core.http.HttpResponse;
+import org.apiguardian.api.API;
+
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.Executor;
+import java.util.function.Consumer;
+
+import static org.apiguardian.api.API.Status.INTERNAL;
+
+@API(since = "0.x", status = INTERNAL)
+public class ClientHttpExecutorImpl<R> extends HttpExecutorImpl<R> {
+
+    public ClientHttpExecutorImpl(final ClientConfiguration configuration, final HttpProvider<R> provider) {
+        super(Assert.requireNonNull(configuration, "configuration").getBackgroundExecutor(), configuration.getUiExecutor(), provider);
+    }
+
+}
